api/requirements.txt
fastapi
uvicorn
pika
pandas
scikit-learn
joblib
nltk
psycopg2-binary